requirements.txt
accelerate==0.34.2
aiohttp==3.10.10
anthropic==0.52.0
backoff==2.2.1
colorama==0.4.6
datasets==2.19.1

exllamav2==0.3.0
fastapi==0.115.12
flash_attn==2.7.4.post1
#gradio==5.31.0
hdbscan==0.8.40
httpx==0.28.1
huggingface_hub==0.26.1
intel_extension_for_pytorch==2.7.0
jsonlines==4.0.0
#kaleido==0.2.1.post1
#llama==0.1.1
#numpy==2.2.6
openai==1.82.0
opencc==1.1.9
packaging==25.0
pandas==2.2.3
parse==1.20.2
peft==0.13.2
plotly==6.1.2
polyglot==16.7.4
prompt_toolkit==3.0.48
psutil==6.1.0
pycld2==0.42
pydantic==2.11.5
pytz==2024.2
#PyYAML==5.4.1
PyYAML==6.0.2
ray==2.38.0
Requests==2.32.3
rich==14.0.0
rwkv==0.8.29
safetensors==0.5.3
scienceworld==1.1.3
scikit_learn==1.2.2
sentence_transformers==3.2.1
shortuuid==1.0.13
tiktoken==0.9.0
#torch==2.5.1+cu121
#torch_npu==2.5.1
tqdm==4.66.5
transformers==4.50.1
trl==0.7.7
uvicorn==0.34.2
vertexai==1.71.1
vllm==0.6.4.post1
wandb==0.19.11
xformers==0.0.28.post3
